+++ CHANGELOG.md
@@ -6,6 +6,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.4] - 15 February 2021
+### Fixed
+- Log all exceptions to NextGenImages logger
+
+## [0.10.3] - 12 February 2021
+### Fixed
+- Make sure configurable images don't throw exception
+
+## [0.10.2] - 12 February 2021
+### Fixed
+- Make sure wrong gallery images don't throw exception
+
+## [0.10.1] - 11 February 2021
+### Fixed
+- Fix JS error
+- Rewrite into synchronous way of detecting WebP
+
+
 ## [0.10.0] - 28 January 2021
 ### Added
 - Improved support on product detail page (configurable swatches + fotorama)